Move methods from Client to Backend for pandas based backends (#3030)
