+++ ibis/backends/csv/__init__.py
@@ -1,10 +1,9 @@
 import pandas as pd
 import toolz
-from pkg_resources import parse_version
 
 import ibis.expr.operations as ops
 import ibis.expr.schema as sch
-from ibis.backends.base.file import BaseFileBackend, FileClient
+from ibis.backends.base.file import BaseFileBackend
 from ibis.backends.pandas.core import execute, execute_node, pre_execute
 from ibis.backends.pandas.execution.selection import physical_tables
 from ibis.expr.scope import Scope
@@ -28,7 +27,11 @@ class CSVTable(ops.DatabaseTable):
         self.read_csv_kwargs = kwargs
 
 
-class CSVClient(FileClient):
+class Backend(BaseFileBackend):
+    name = 'csv'
+    extension = 'csv'
+    table_class = CSVTable
+
     def insert(self, path, expr, index=False, **kwargs):
         path = self.root / path
         data = execute(expr)
@@ -56,15 +59,11 @@ class CSVClient(FileClient):
 
         return table
 
-    @property
-    def version(self):
-        return parse_version(pd.__version__)
 
-
-@pre_execute.register(ops.Selection, CSVClient)
+@pre_execute.register(ops.Selection, Backend)
 def csv_pre_execute_selection(
     op: ops.Node,
-    client: CSVClient,
+    client: Backend,
     scope: Scope,
     timecontext: TimeContext = None,
     **kwargs,
@@ -95,14 +94,7 @@ def csv_pre_execute_selection(
     return ops
 
 
-class Backend(BaseFileBackend):
-    name = 'csv'
-    extension = 'csv'
-    table_class = CSVTable
-    client_class = CSVClient
-
-
-@execute_node.register(Backend.table_class, CSVClient)
+@execute_node.register(Backend.table_class, Backend)
 def csv_read_table(op, client, scope, **kwargs):
     path = client.dictionary[op.name]
     df = _read_csv(path, schema=op.schema, header=0, **op.read_csv_kwargs)
