+++ ibis/backends/dask/execution/generic.py
@@ -51,7 +51,8 @@ from ibis.backends.pandas.execution.generic import (
     execute_table_column_df_or_df_groupby,
 )
 
-from ..client import DaskClient, DaskTable
+from .. import Backend as DaskBackend
+from ..client import DaskTable
 from ..core import execute
 from ..dispatch import execute_node
 from .util import (
@@ -141,7 +142,7 @@ DASK_DISPATCH_TYPES: TypeRegistrationDict = {
 
 register_types_to_dispatcher(execute_node, DASK_DISPATCH_TYPES)
 
-execute_node.register(DaskTable, DaskClient)(execute_database_table_client)
+execute_node.register(DaskTable, DaskBackend)(execute_database_table_client)
 
 
 @execute_node.register(ops.ValueList, collections.abc.Sequence)
