+++ ibis/backends/pandas/tests/execution/test_cast.py
@@ -8,8 +8,7 @@ from pytest import param
 
 import ibis
 import ibis.expr.datatypes as dt  # noqa: E402
-
-from ... import execute
+from ibis.backends.pandas.execution import execute
 
 
 @pytest.mark.parametrize('from_', ['plain_float64', 'plain_int64'])
