+++ ibis/backends/pandas/tests/execution/test_operations.py
@@ -9,8 +9,8 @@ import pytest
 
 import ibis
 import ibis.expr.datatypes as dt
-
-from ... import Backend, execute
+from ibis.backends.pandas import Backend
+from ibis.backends.pandas.execution import execute
 
 
 def test_table_column(t, df):
